requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
tqdm>=4.66.0
transformers>=4.38.0
pytorch-lightning>=2.0.0
pillow>=10.0.0
gradio>=4.0.0
pytest>=8.0.0
